feat(ncubed): query root route and dispatch message
This code is subtly bnroken right now. When receiving the HTTP request I
dispatch a message over a channel and query the local Ncube database.
This works but only for a single request. On the second request the
application panics. This is somehow related to the pool. Right now I'm
not sure how to handle this.