feat(ncube): start ncubed in the background when starting Ncube.
`ncubed` was refactored to act more like a library. The approach taken
right now is that in Ncube I spawn a thread in which I create a new
threaded tokio runtime and spawn the Ncube background process.

The Ncube UI is started in the regular main thread of Ncube.

Fixes #17