fix(ncubed): normalized sql include path in config store
CI threw an error related to the include paths of the sql snippets for
the config store. See
[here](https://github.com/critocrito/ncube/runs/740537917) for more
details. The error did not show when building locally. For unknown
reason the `src/rs/ncubed/src/stores/config.rs` uses a different path
when importing locally. I hope by setting a link to the `sql` directory
in the config root I can avoid this behavior locally and on CI.