chore(release): release version 0.5.0
This release focuses on a server release for Ncube. The `ncubed` and
`ncubectl` binaries are important parts of Ncube from now. This release
sees a newDebian package target which is aimed at server only
installations. To support server operations the `ncubectl` program can
be used to create workspaces and accounts, and other aspects of the
server operations of Ncube. The Ncube client can link to remote
workspaces provided by the Ncube daemon.