refactor(ui): feature parity between typescript and clojurescript
The typescript version does a little moreeven. It validates the input of
all forms and the POST requests to the HTTP API. The clojurescript
actually renders sources as a table, which the typescript version
doesn't do yet. It only renders sources as a list. This will be fixed
when building the data table proper.