chore(release): release version 0.7.0
The biggest feature of this release is the ability to run data processes
from within Ncube. Currently, those data processes can preserve videos
from Youtube and tweets from Twitter. The data details page shows the
full unit of data with all media attached.

The Discovery browser extension got a complete redesign and builds now
for Firefox and Chrome.

There are many small improvements that can be found in the Changelog.