chore(release): release version 0.8.0
The main feature of this release is the ability to create investigations
and send segments to an investigation for verification. The
investigation itself is rendered as a kanban board for each segment.
Units of data are dragged as cards between the different states. The
different states of an investigation are described in a methodology
format. While right now only a single methodology comes with Ncube out
ofthe box, it would be trivial to allow new methodologies to be
uploaded.

The search query grammar was extended to limit searches by tags and
sources. A new syntax was added to support such limits. The search
grammar itself is described as a PEG grammar.

The number of stats that are *live* was massively increased. Almost
everywhere static stats were replaced with real requests.

Every screen of Ncube contains some description text with further links
pointing to the website. Those links open in an external browser. Some
of the copy texts are still missing or could use some refactoring.

The preservation was improved with enhanced support on the Sugarcube
side of things.