feat(ui): delete segments using the UI
It was always possible to delete segments using the HTTP API. But there
was no facility in the UI to do so. Further, the deletion didn't really
work, since we violated a foreign key constraint. The UI got a new
button to delete segments and the backend deletes verifications and
annotations along with the referenced segment.

refs #53