refactor(ncubed): change tasks into structs with named fields
This allowed me to remove the manual serialization of tasks since the
field names are named and I can use the derive macro instead. I also
found myself not understanding the fields and their meanings. This meant
that I had to look inside the task runner dispatch what the fields mean.