feat(ncubed): send task update messages to the clients
This implementation is rather rudimentary. I probably should extract the
PubSub mechanism with the clients into a dedicated crate and wrap the
messaging into an easy to use API. Right now it is quite verbose to
send messages to the client. Further, clients are not filtered by
subscription topic and the messages aren't really typed. It's a step in
the right direction though.

refs #55