refactor(ncubed): push progress messages inside of task
Tasks are separate functions that are called by the task runner. Until
now, progress messages were pushed from the task runner. With this
change progress messages can be pushed from inside the task. This allows
not only for a fine grained control of messaging, but in the future I
can push stdout log messages as progress messages as well.

refs #55