refactor: change frontend application structure
This is a huge commit. It restructures the elements of the UI
application and tries to introduce consistent handling of state-charts,
components and names. All state-charts are collected in the `views`
directory. The complete state of the application is managed within those
state-charts. The state-charts make use of actors as well. This makes it
easier how different elements interact.