fix: support HTTP Range header for file downloads
Video playback is broken (definitely under Safari) because downloading
videos didn't support the `RANGE` header. This is a requirement though.
This commit adds support for `RANGE` requests, but doesn't yet support
any of the other conditional HTTP headers. It might be interesting to
support the various `IF-MODIFIED*` headers to allow for resuming
downloads. I guess this will be more important once I add support for
bundle download and remote workspace downloads.

refs #48